+++ src/d_player.h
@@ -232,6 +232,7 @@ typedef struct player_s
     int             itemspickedup_ammo_shells;
     int             itemspickedup_armor;
     int             itemspickedup_health;
+    int             infightcount;
 } player_t;
 
 //