+++ releasenotes.md
@@ -14,6 +14,7 @@
 * The fading of player messages onto and off of the screen is now smoother in some instances when the `fade` CVAR is `on`.
 * A fade transition will now be applied when changing the `r_brightmaps` CVAR in the console and when the `fade` CVAR is `on`.
 * Corpses and dropped items will no longer perpetually slide back and forth in some rare instances.
+* The number of monsters killed due to infighting are now displayed by the `playerstats` CCMD.
 
 ![](https://github.com/bradharding/www.doomretro.com/raw/master/wiki/bigdivider.png)